refactor(registration-react): Separate out distributionTypes helper from the FormDistribution
